+++ actix-router/CHANGES.md
@@ -2,8 +2,10 @@
 
 ## Unreleased - 2021-xx-xx
 * When matching URL parameters, `%25` is kept in the percent-encoded form - no longer decoded to `%`. [#357]
+* Fixed a bug where the `Path` extractor returns unsafe malformed string due to malformed URL. [#359]
 
 [#357]: https://github.com/actix/actix-net/pull/357
+[#359]: https://github.com/actix/actix-net/pull/359
 
 
 ## 0.2.7 - 2021-02-06